06-final-project/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
redis
openai
